mobile-q-and-a/requirements.txt
# requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-huggingface
openai
tiktoken
requests
python-dotenv
gradio
httpx
faiss-cpu==1.7.4
sentence-transformers==2.5.1
